requirements.txt
Babel==1.3
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Pygments==2.0.2
Sphinx==1.3.1
alabaster==0.7.3
argh==0.26.1
bitarray==0.8.1
certifi==14.05.14
decorator==3.4.2
docutils==0.12
gnureadline==6.3.3
ipython==3.1.0
livereload==2.3.2
matplotlib==1.4.3
mmh3==2.3
networkx==1.9.1
nose==1.3.6
numpy==1.9.2
pathtools==0.1.2
pycrypto==2.6.1
pydotplus==2.0.2
pygraphviz==1.3rc2
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.2
simpy==3.0.7
six==1.9.0
snowballstemmer==1.2.0
sphinx-autobuild==0.5.2
sphinx-rtd-theme==0.1.7
sympy==0.7.6
tornado==4.1
watchdog==0.8.3
